Signed-off-by: Bram Dufour <britaj@bancolombia.com.co>
Added HasOUValue func that checks if an OU with a specified value is present in the client's X509 cert without impacting the clientIdentity interface.

This way granular access control in the chaincode can also be done based on OU values instead of only on attributes from the custom extra extension in the cert that is not supported by external CA services.
To not impact the interface clientIdentity or affect current users of the package, but still be able to add HasOUValue to the type returned by New in cid.go the following refactoring was done: changed clientIdentityImpl to ClientID, changed return type of New func to *ClientID.

FABCG-12: https://jira.hyperledger.org/browse/FABCG-12

Unit tests added for HasOUValue + existing tests were run and passed.