Add gh-actions workflow
Update the two lint and the one failing assertion that prevent the build from completing
The fixes for the assertion shoul dbe a temporary solution as I believe we need to change the assertion library.
To the one use by the main fabric repo - as that seems to be able to cope with equality of TLSConfig structures.

Current one here doesn't cope with objects with functions.

Signed-off-by: Matthew B White <whitemat@uk.ibm.com>