Refactor write batching (#143)
* Refactor write batching

Avoid holding transaction-specific state in the handler, which must then
be protected by multiple locks. This is error prone and is a potential
performance bottleneck. Instead, extract batch writes to a separate
struct for clarity, and maintain the batch write state in the stub,
which is unique to a specific transaction so requires no locking.

The stub (and associated batch write state) is also discarded after a
transaction invocation so there is no risk of memory leaks due to
redundant batch write state being left in the shared handler.

Signed-off-by: Mark S. Lewis <Mark.S.Lewis@outlook.com>

* Fix issues raised in review

- Check whether batching is supported when starting a batch.
- Better keying of batched writes.

Signed-off-by: Mark S. Lewis <Mark.S.Lewis@outlook.com>

* Correct types for batch records

Signed-off-by: Mark S. Lewis <Mark.S.Lewis@outlook.com>

* Address review comments

- Preallocate writes slice.

Signed-off-by: Mark S. Lewis <Mark.S.Lewis@outlook.com>

---------

Signed-off-by: Mark S. Lewis <Mark.S.Lewis@outlook.com>